resolved more warnings

+++ CollectGame/UI.cpp
@@ -22,7 +22,7 @@ namespace Tmpl8
 
 	void UI::initLevelTxt(Surface* surface)
 	{
-		levelTxtAnimation.init(1, "assets/temp_Level_txt.png", levelTxtLoc.x, levelTxtLoc.y, surface);
+		levelTxtAnimation.init(1, "assets/temp_Level_txt.png", static_cast<int>(levelTxtLoc.x), static_cast<int>(levelTxtLoc.y), surface);
 	}
 	void UI::initLevelCounter(Surface* surface, int levelNum)
 	{
@@ -36,19 +36,19 @@ namespace Tmpl8
 		std::strcpy(filenameBuffer, filename.c_str());
 		//until here
 
-		levelCounterAnimation.init(1, filenameBuffer, levelCounterLoc.x, levelCounterLoc.y, surface);
+		levelCounterAnimation.init(1, filenameBuffer, static_cast<int>(levelCounterLoc.x), static_cast<int>(levelCounterLoc.y), surface);
 	}
 
 	void UI::updateLevelCounter(Surface* surface, int levelNum)
 	{
-		levelTxtAnimation.update(1, levelTxtLoc.x, levelTxtLoc.y, 120, 30, surface);
+		levelTxtAnimation.update(1, static_cast<int>(levelTxtLoc.x), static_cast<int>(levelTxtLoc.y), 120, 30, surface);
 
 		if (currentLevelNum != levelNum)
 		{
 			initLevelCounter(surface, levelNum);
 		}
 		
-		levelCounterAnimation.update(1, levelCounterLoc.x, levelCounterLoc.y, 30, 30, surface);
+		levelCounterAnimation.update(1, static_cast<int>(levelCounterLoc.x), static_cast<int>(levelCounterLoc.y), 30, 30, surface);
 	}
 
 	void UI::initHpCounter(Surface* surface, int hp)
@@ -56,7 +56,7 @@ namespace Tmpl8
 		for (int i = 0; i < hp; i++)
 		{
 			heartAnimVec.push_back(new animation);
-			heartAnimVec.back()->init(1, "assets/temp_Heart.png", heartLoc.x - 40 * i, heartLoc.y, surface);
+			heartAnimVec.back()->init(1, "assets/temp_Heart.png", static_cast<int>(heartLoc.x - 40 * i), static_cast<int>(heartLoc.y), surface);
 		}
 	}
 	void UI::updateHpCounter(int hp, Surface* surface)
